requirements.txt
fastapi
aiogram==2.25
python-dotenv
torch
diffusers
Pillow
redis
